feat: change blocking component idea & code (WIP)
TODO: bookmark should show only in the book being updated.
For this, I plan to add a state to `<Book>`, that will became
`true` when the book `onUpdateBook` happens and `false` when the
updating finishes. With this state, it will control/add a proper
CSS class to show/hide the related bookmark.